requirements.txt
 streamlit==1.44.1
pdfplumber>=0.7.0
groq>=0.3.0
openai>=0.27.0
anthropic>=0.5.0
google-generativeai>=0.2.0
python-dotenv>=0.19.0
pytest>=7.0.0
